:minibus: Move openExternally to common helpers
