+++ lib/tooltip/message-legacy.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import React from 'react'
-import { openExternally } from './helpers'
+import { openExternally } from '../helpers'
 import type TooltipDelegate from './delegate'
 import type { MessageLegacy } from '../types'
