+++ lib/tooltip/message.js
@@ -3,8 +3,7 @@
 import React from 'react'
 import marked from 'marked'
 
-import { visitMessage } from '../helpers'
-import { openExternally } from './helpers'
+import { visitMessage, openExternally } from '../helpers'
 import type TooltipDelegate from './delegate'
 import type { Message } from '../types'
